Added batch mode
Batch mode was needed for my use case, and it required implementing an interface. Also was tired of the issue with it responding with unhelpful error messages, so recoded that to the specs in documentation. To do that it must use non Basic handler for http request (always receive body alongside header status code) as well. 
Also added some helper methods to the primary message for headers and attachments. 

HAS NOT BEEN TESTED YET.